attendance/requirements.txt
Django==5.2.10
djangorestframework==3.16.1
gunicorn==21.2.0
whitenoise==6.6.0

